download/unpack mechanics

+++ solr/build-1-get-sources.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
+force=0 
+
+echo "Where files EXIST -- please remove manually to refresh."
+
+echo
 echo
 echo "Download NGA Geonames - World"
 echo "==============================="
@@ -7,18 +12,22 @@ TEST=tmp/$FILE
 if [ !  -f "$TEST" ]; then 
     curl -k  https://geonames.nga.mil/geonames/GNSData/fc_files/$FILE -o $TEST
 else
-    echo "EXISTS - not overwriting"
+    echo "EXISTS $TEST"
 fi
 
-if [ -f "$TEST" ]; then
+TARGET=./tmp/Whole_World.txt
+if [ -f "$TEST" -a ! -e $TARGET ]; then
   echo "Unpacking $TEST"
   # On Mac:  7zz;  On Ubuntu: 7za
   UNZIP=7zz
-  TARGET=./tmp/Whole_World.txt
-  (cd ./tmp && $UNZIP e -so  $TEST | grep -v "The geographic names in this database" > Whole_World.txt)
+  rm $TARGET
+  $UNZIP e -so  $TEST | grep -v "The geographic names in this database" > $TARGET
   echo "NGA Geonames file is at $TARGET"
+else
+    echo "EXISTS $TARGET ?"
 fi
 
+echo
 echo
 echo "Download USGS National File"
 echo "==============================="
@@ -28,20 +37,25 @@ TEST=tmp/$FILE
 if [ !  -f "$TEST" ]; then 
     curl -k "https://geonames.usgs.gov/docs/stategaz/$FILE" -o $TEST
 else
-    echo "EXISTS - not overwriting"
+    echo "EXISTS $TEST"
 fi
 
-if [ -f "$TEST" ]; then
+TARGET=./tmp/NationalFile.txt
+if [ -f "$TEST" -a ! -e $TARGET ]; then
   echo "Unpacking $TEST"
   UNZIP=unzip
   $UNZIP -d ./tmp  $TEST
-  TARGET=./tmp/NationalFile.txt
   mv ./tmp/NationalFile_${USGS_PRODDATE}* $TARGET
 
   echo "USGS file is at $TARGET"
+else
+  echo "EXISTS $TARGET ?"
 fi
 
 
+echo
+echo
+echo "==============================="
 echo "Download HumData Exchange - PAK"
 echo "==============================="
 FILE=pak_adm_wfp_20220909_shp.zip
@@ -55,6 +69,7 @@ if [ !  -f "$TEST" ]; then
   read ans
   if [ "$ans" = "y" -a -f "$TEST" ]; then
     echo "Continuing ... "
+    unzip -d tmp/pak_adm_wfp $TEST
   else
     echo "File not found"
   fi
